EmergencyResponse/requirements.txt
flask==2.0.1
twilio==7.17.0
opencv-python==4.8.0.74
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
geopy==2.3.0
pytz==2023.3
schedule==1.2.0
python-socketio==5.8.0
flask-socketio==5.3.4